more cursor refactoring
